+++ recipes/enable-enhanced-networking.rb
@@ -6,7 +6,7 @@
 include_recipe "mh-opsworks-recipes::install-awscli"
 install_package('dkms')
 bucket_name = get_shared_asset_bucket_name
-dkms_version = node.fetch(:ixgbevf_version, "2.16.1")
+dkms_version = node.fetch(:ixgbevf_version, "2.16.4")
 
 cookbook_file 'enable_enhanced_networking.sh' do
   path "/usr/local/bin/enable_enhanced_networking.sh"